backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
chromadb
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
websockets
